Fixed folder for enabled customization not being created (#58)
Previously, the editor would check if both the customization and enabled folders are created before handling file-based customizations. The better way of doing that is only check if the customizations folder exists (because it should) and create the enabled folder thereafter if it is not created (which it may not be).