feat: configure services for dependency injection
Configures the logger, Notifier, Utilities, and MainWindow as services in the dependency injection container. This allows them to be automatically provided when specified in a constructor. However, since the BtnInstallContentConverter requires a parameterless constructor, it had to new up its dependencies itself. These should be moved out in the future.